# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gerenciamento de eventos fotográficos",
        "title": "Photo Event API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/api/v1",
    "paths": {
...
        "/auth/login": {
            "post": {
                "description": "Realiza login de um usuário no sistema",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login do usuário",
                "parameters": [
                    {
                        "description": "Credenciais de login",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Realiza logout do usuário (não implementado)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Logout",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Renova o token de acesso (não implementado)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh Token",
                "responses": {
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Registra um novo usuário no sistema",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Registrar usuário",
                "parameters": [
                    {
                        "description": "Dados de registro",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/schemas.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/users": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Show all users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Show users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ListUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_photo-event-backend_handlers_users_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_photo-event-backend_handlers_users_handler.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create an user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Create an user",
                "parameters": [
                    {
                        "description": "Create User Params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_photo-event-backend_handlers_users_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_photo-event-backend_handlers_users_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_isaiaspereira307_photo-event-backend_handlers_users_handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error_code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handlers.CreateUserRequest": {
            "type": "object",
            "required": [
                "email",
                "first_name",
                "last_name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "handlers.CreateUserResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/schemas.UserResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handlers.ListUserResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schemas.UserResponse"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "schemas.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Erro de validação"
                },
                "message": {
                    "type": "string",
                    "example": "Campo obrigatório não informado"
                }
            }
        },
        "schemas.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "senha"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "joao@exemplo.com"
                },
                "senha": {
                    "type": "string",
                    "example": "senha123"
                }
            }
        },
        "schemas.LoginResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Login realizado com sucesso"
                },
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "user": {
                    "$ref": "#/definitions/schemas.User"
                }
            }
        },
        "schemas.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "nome",
                "senha"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "joao@exemplo.com"
                },
                "nome": {
                    "type": "string",
                    "minLength": 2,
                    "example": "João Silva"
                },
                "senha": {
                    "type": "string",
                    "minLength": 6,
                    "example": "senha123"
                },
                "tipo": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/schemas.UserRole"
                        }
                    ],
                    "example": "usuario"
                }
            }
        },
        "schemas.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Operação realizada com sucesso"
                }
            }
        },
...
        "schemas.User": {
            "type": "object",
            "properties": {
                "biografia": {
                    "description": "Campos específicos para fotógrafos",
                    "type": "string"
                },
                "creditos": {
                    "type": "number"
                },
                "data_cadastro": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "telefone": {
                    "type": "string"
                },
                "tipo": {
                    "$ref": "#/definitions/schemas.UserRole"
                },
                "ultimo_acesso": {
                    "type": "string"
                },
                "url_capa": {
                    "type": "string"
                },
                "url_perfil": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        },
        "schemas.UserResponse": {
            "type": "object",
            "properties": {
                "biografia": {
                    "type": "string",
                    "example": "Fotógrafo profissional"
                },
                "creditos": {
                    "type": "number",
                    "example": 100.5
                },
                "data_cadastro": {
                    "type": "string",
                    "example": "2025-08-19T14:00:00Z"
                },
                "email": {
                    "type": "string",
                    "example": "joao@exemplo.com"
                },
                "id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "nome": {
                    "type": "string",
                    "example": "João Silva"
                },
                "telefone": {
                    "type": "string",
                    "example": "(11) 99999-9999"
                },
                "tipo": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/schemas.UserRole"
                        }
                    ],
                    "example": "usuario"
                },
                "ultimo_acesso": {
                    "type": "string",
                    "example": "2025-08-19T15:00:00Z"
                },
                "url_capa": {
                    "type": "string",
                    "example": "https://exemplo.com/capa.jpg"
                },
                "url_perfil": {
                    "type": "string",
                    "example": "https://exemplo.com/perfil.jpg"
                },
                "website": {
                    "type": "string",
                    "example": "https://joaosilva.com"
                }
            }
        },
        "schemas.UserRole": {
            "type": "string",
            "enum": [
                "admin",
                "fotografo",
                "usuario"
            ],
            "x-enum-comments": {
                "RoleAdmin": "Administrador do sistema",
                "RoleFotografo": "Fotógrafo profissional",
                "RoleUsuario": "Usuário comum"
            },
            "x-enum-varnames": [
                "RoleAdmin",
                "RoleFotografo",
                "RoleUsuario"
            ]
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Bearer token para autenticação",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}